+++ NEWS.md
@@ -1,6 +1,14 @@
 History of LombScargle.jl
 ==========================
 
+v0.2.0 (2016-??-??)
+-------------------
+
+### New Features ###
+
+* New functions: `findmaxpower`, `prob`, `probinv`, `fap`, `fapinv`.
+* New optional keyword for `lombscargle` function: `noise_level`.
+
 v0.0.1 (2016-07-16)
 -------------------